enable multiple click-outside targets
This allows EzField[type=select] to render the options in a different
DOM hierarchy than the combobox input element (i.e. at the bottom of the
page).

This facilitates rendering EzField[type=select] options in a portal in
order to escape any `overflow: hidden` containers that would otherwise
clip the contents of the options flyout.