fix: EzLayout breaking stack layouts in IE
The `flex-basis: 0` value is causing content to lose its intrinsic width and in IE the flex items aren't filling the available space, thereby causing text content to wrap and stack instead of fill the available space.

We want stack items to fill the available width, so setting `width: 100%` does the trick. We then need to reset the width back to auto on responsive layouts.