fix: update API for disabling stitches CSS insertion
When we're running inside Jest, the document object is provided by JSDOM, however since JSDOM doesn't fully support CSSOM, it cannot generate styles correctly. Internally, Stitches provides a fallback for server environments that doesn't use CSSOM, but it doesn't kick in if `document` is present.

Explicitly setting the root node within stitches to `null` allows the fallback logic in stitches to kick in.